requirements.txt
azure-ai-inference
aiohttp
ipykernel
python-dotenv
pandas
openpyxl